add sphinx-argparse documentation
