+++ docs/requirements.txt
@@ -1,0 +1,6 @@
+# docs
+furo
+sphinx
+sphinx-argparse
+myst-parser
+mdit-py-plugins>=0.3.0
