+++ docs/source/cli.md
@@ -1,0 +1,9 @@
+Command Line
+============
+
+```{eval-rst}
+.. argparse::
+   :module: grayskull.main
+   :func: build_parser
+   :prog: grayskull
+```
